expand assembly bundle, program now top level
